all mcni tests passed
git-svn-id: svn://danse.us/MCViNE/trunk@873 bd03dbb6-ca95-436e-b6ad-03a95796df24

+++ packages/mcni/tests/mcni/pyre_components/ndmonitor/ndmonitor_mpi_TestCase.py
@@ -20,9 +20,9 @@ class TestCase(unittest.TestCase):
 
     def test3(self):
         outdir = 'out-testmpi'
-        shutil.rmtree(outdir)
+        if os.path.exists(outdir):
+            shutil.rmtree(outdir)
         cmd = './testmpi -mpirun.nodes=2'
-        import os
         if os.system(cmd):
             raise RuntimeError, "%r failed" % cmd
